# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления токенами и логами ботов",
        "title": "Logging API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/bots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех ботов (требуется админский токен)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bots"
                ],
                "summary": "Получить всех ботов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Bot"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт нового бота/робота (требуется админский токен)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bots"
                ],
                "summary": "Создать бота",
                "parameters": [
                    {
                        "description": "Данные бота",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bot_handler.CreateBotRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Bot"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/bots/{bot_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает информацию о боте по ID (требуется админский токен)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bots"
                ],
                "summary": "Получить бота",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID бота (UUID)",
                        "name": "bot_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Bot"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет данные бота (требуется админский токен)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bots"
                ],
                "summary": "Обновить бота",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID бота (UUID)",
                        "name": "bot_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновлённые данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bot_handler.UpdateBotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Bot"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Полностью удаляет бота из базы данных (требуется админский токен)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bots"
                ],
                "summary": "Удалить бота",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID бота (UUID)",
                        "name": "bot_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bot_handler.CreateBotRequest": {
            "type": "object",
            "required": [
                "bot_type",
                "code",
                "language",
                "name"
            ],
            "properties": {
                "bot_type": {
                    "type": "string",
                    "enum": [
                        "AI",
                        "Backend",
                        "Frontend",
                        "Robot"
                    ],
                    "example": "Backend"
                },
                "code": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "example": "BOT_001"
                },
                "description": {
                    "type": "string",
                    "example": "Бот для обработки сообщений"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "language": {
                    "type": "string",
                    "enum": [
                        "Python",
                        "Go",
                        "N8N",
                        "PIX",
                        "JS",
                        "C",
                        "Other"
                    ],
                    "example": "Python"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Telegram Bot"
                },
                "owner_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "telegram",
                        "bot"
                    ]
                }
            }
        },
        "bot_handler.UpdateBotRequest": {
            "type": "object",
            "properties": {
                "bot_type": {
                    "type": "string",
                    "enum": [
                        "AI",
                        "Backend",
                        "Frontend",
                        "Robot"
                    ],
                    "example": "AI"
                },
                "code": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "example": "BOT_002"
                },
                "description": {
                    "type": "string",
                    "example": "Обновлённое описание"
                },
                "is_active": {
                    "type": "boolean",
                    "example": false
                },
                "language": {
                    "type": "string",
                    "enum": [
                        "Python",
                        "Go",
                        "N8N",
                        "PIX",
                        "JS",
                        "C",
                        "Other"
                    ],
                    "example": "Go"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Discord Bot"
                },
                "owner_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "discord",
                        "ai"
                    ]
                }
            }
        },
...
        "models.Bot": {
            "type": "object",
            "required": [
                "bot_type",
                "code",
                "language",
                "name"
            ],
            "properties": {
                "bot_type": {
                    "description": "Тип бота (AI, Backend, Frontend, Robot)",
                    "type": "string",
                    "enum": [
                        "AI",
                        "Backend",
                        "Frontend",
                        "Robot"
                    ],
                    "example": "Backend"
                },
                "code": {
                    "description": "Уникальный код бота для идентификации",
                    "type": "string",
                    "example": "BOT_001"
                },
                "created_at": {
                    "description": "Дата и время создания записи",
                    "type": "string",
                    "example": "2023-01-15T12:00:00Z"
                },
                "description": {
                    "description": "Подробное описание функционала бота",
                    "type": "string",
                    "example": "Бот для обработки сообщений"
                },
                "id": {
                    "description": "Уникальный идентификатор бота (UUID)",
                    "type": "string",
                    "format": "uuid",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "is_active": {
                    "description": "Активен ли бот",
                    "type": "boolean",
                    "example": true
                },
                "language": {
                    "description": "Основной язык или технология реализации",
                    "type": "string",
                    "enum": [
                        "Python",
                        "Go",
                        "N8N",
                        "PIX",
                        "JS",
                        "C",
                        "Other"
                    ],
                    "example": "Python"
                },
                "name": {
                    "description": "Название бота",
                    "type": "string",
                    "example": "Telegram Bot"
                },
                "owner_id": {
                    "description": "ID владельца бота",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "tags": {
                    "description": "Массив тегов для классификации",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "telegram",
                        "automation"
                    ]
                },
                "updated_at": {
                    "description": "Дата и время последнего обновления",
                    "type": "string",
                    "example": "2023-01-15T12:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}